Some changes. Version 7.0.0 released to people.
CHANGED: Size of MainFrame.
FIXED: getUI() not found.
FIXED: An error with Minecraft load system.
FIXED: An error with META-INF in MC versions newer than 1.6.
FIXED: An error with DownloadJob.
FIXED: An error with deleteDirectory which throws NullPointerException.
FIXED: An error with the new load system. JFrame didn't load correctly.